% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mpg.R
\name{compare_mpg}
\alias{compare_mpg}
\title{Compare MPG by Cylinder Count}
\usage{
compare_mpg(data)
}
\arguments{
\item{data}{A dataframe with `mpg` and `cyl` columns.}
}
\value{
A ggplot boxplot.
}
\description{
Compare MPG by Cylinder Count
}
