% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Statistics for a Numeric Vector}
\usage{
summary_stats(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of summary statistics.
}
\description{
Summary Statistics for a Numeric Vector
}
