% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_speed_cars.R
\name{top_speed_cars}
\alias{top_speed_cars}
\title{Compare MPG by Cylinder Count}
\usage{
top_speed_cars(data, top_n = 5)
}
\arguments{
\item{data}{A dataframe with `mpg` and `cyl` columns.}
}
\value{
A ggplot boxplot.
}
\description{
Compare MPG by Cylinder Count
}
